requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
langchain==0.0.350
langchain-openai==0.0.2
openai==1.3.7
